Created FamilyCare
A family care tab that doesn't have too much on it. It was one of the least concerning features of the app but visually it has 4 location markers for the mom, dad, friend, and the user. There is a sliderscroll that has all the names and a button for more info on the whereabouts of friends and family.